services/behavior-analysis/requirements.txt
# Behavior Analysis Microservice
# Real-time face/hand/voice analysis using OpenCV + MediaPipe

fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
opencv-python-headless==4.9.0.80
mediapipe==0.10.11
numpy==1.26.4
scipy==1.12.0
librosa==0.10.1
soundfile==0.12.1
pydantic==2.6.3
python-dotenv==1.0.1
httpx==0.27.0
cryptography==42.0.5
redis==5.0.1
aiofiles==23.2.1
orjson==3.9.15
Pillow==10.2.0
